pyproject.toml
[tool.poetry]
name = "juju"
version = "0.1.0"
description = ""
authors = ["Vahid Al <thevahidal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.2"
django-json-widget = "^2.0.1"


[tool.poetry.group.dev.dependencies]
django = "^5.1"
django-stubs = "^5.0.4"
djangorestframework = "^3.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
